Move test_write_changed_file() into test_write.py
